requirements.txt
protobuf==6.33.1
pytest==9.0.1
python-dotenv==1.2.1
